+++ src/shared/gems_view_component.rb
@@ -4,15 +4,21 @@ initializer "view_component.rb", <<-CODE
   Rails.application.config.view_component.generate.sidecar = true
 CODE
 
+# Create or overwrite inflections.rb initializer
+inflections_file = "config/initializers/inflections.rb"
+remove_file inflections_file if File.exist?(inflections_file)
+initializer "inflections.rb", <<-CODE
+ActiveSupport::Inflector.inflections(:en) do |inflect|
+  inflect.acronym "UI"
+end
+CODE
+
 copy_file "app/helpers/view_component_helper.rb", force: true
 copy_file "app/views/layouts/component_preview.html.erb", force: true
 
 directory "app/components", force: true
 directory "spec/components", force: true
 
-layout_file = "app/views/layouts/application.html.erb"
-if File.exist?(layout_file)
-  gsub_file layout_file,
-            "<%= yield %>",
-            "<%= component \"ui/alert\", flash: flash %>\n    <%= yield %>"
-end
+# Copy view component support files for RSpec
+copy_file "spec/support/view_component.rb", force: true
+copy_file "spec/support/capybara.rb", force: true