+++ variants/shared/app/controllers/landing_controller.rb
@@ -1,0 +1,9 @@
+class LandingController < ApplicationController
+  layout "landing"
+
+  def home
+  end
+
+  def about
+  end
+end