+++ variants/shared/app/controllers/pages_controller.rb
@@ -1,0 +1,9 @@
+class PagesController < ApplicationController
+  layout "plain"
+
+  def terms
+  end
+
+  def privacy
+  end
+end