+++ variants/shared/spec/requests/landing_spec.rb
@@ -1,0 +1,17 @@
+require "rails_helper"
+
+RSpec.describe "Landing", type: :request do
+  describe "GET /home" do
+    it "returns http success" do
+      get "/"
+      expect(response).to have_http_status(:success)
+    end
+  end
+
+  describe "GET /about" do
+    it "returns http success" do
+      get "/about"
+      expect(response).to have_http_status(:success)
+    end
+  end
+end