+++ variants/shared/spec/support/capybara.rb
@@ -1,0 +1,5 @@
+require "capybara/rspec"
+
+RSpec.configure do |config|
+  config.include Capybara::RSpecMatchers, type: :component
+end